% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra.R
\name{obm_computation}
\alias{obm_computation}
\title{Background Computation Function}
\usage{
obm_computation(
  action = "",
  data_files = NULL,
  params = NULL,
  config_file = NULL
)
}
\arguments{
\item{action}{post, get-results, get-status}

\item{data_files}{list of files to sending}

\item{params}{computation control params}

\item{config_file}{is The computation_conf.yml file}
}
\value{
data frame or boolean
}
\description{
This experimental function allows to manage background computation on an OpenBioMaps server.
}
\examples{
\donttest{
# Manage computational packs
results <- obm_computation(action='post',
                           data_files='examples/scripts/bombina.R',
                           params=list(package = 'my_project'),
                           config_file='examples/computation_conf.yml')
}
}
\keyword{computation}
