% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{obm_get_graphql}
\alias{obm_get_graphql}
\title{obm_get_graphql helper function}
\usage{
obm_get_graphql(
  scope,
  control_condition,
  condition,
  token,
  url,
  table,
  retry = TRUE
)
}
\description{
Helper to execute GraphQL query
Not intended for direct user use.
}
\keyword{internal}
