% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{obm_get_rest_v3}
\alias{obm_get_rest_v3}
\title{obm_get_rest_v3 helper function}
\usage{
obm_get_rest_v3(scope, control_condition, condition, token, url, table)
}
\description{
Helper to execute REST API v3 queries (non-GraphQL)
Not intended for direct user use.
}
\keyword{internal}
