% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{obm_parse_control}
\alias{obm_parse_control}
\title{obm_parse_control helper function}
\usage{
obm_parse_control(control_string)
}
\description{
Helper function to parse control string (e.g. "limit=100:0")
Not intended for direct user use.
}
\keyword{internal}
