% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{obm_parse_filters}
\alias{obm_parse_filters}
\title{obm_parse_filters helper function}
\usage{
obm_parse_filters(condition)
}
\description{
Helper to construct GraphQL filter from condition list
Not intended for direct user use.
}
\keyword{internal}
